Decouple widgets from MainWindow
We no longer dig deep into MainWindow to access the trace or other data.
We only use the main window for opening links and the like. To ensure we
don't regress on this we introduce an interface.

Furthermore, we no longer make use of the main window's theme.Window, by
no longer doing layout work in a widget's constructor. Instead, the
first call to Layout runs initialization code with the passed-in
theme.Window. This is in preparation of a future change that improves
the handling of widgets that move between windows.